+++ src/main/java/titan/ccp/aggregation/experimental/teetimebased/SensorHistory.java
@@ -6,12 +6,12 @@ import titan.ccp.model.sensorregistry.MachineSensor;
 
 public interface SensorHistory {
 
-	public long getOrZero(MachineSensor machineSensor);
+	public double getOrZero(MachineSensor machineSensor);
 
 	// public Long getOrNull(MachineSensor machineSensor);
 
-	public void update(MachineSensor machineSensor, long value);
+	public void update(MachineSensor machineSensor, double value);
 
-	public void update(MachineSensor machineSensor, long value, Instant time);
+	public void update(MachineSensor machineSensor, double value, Instant time);
 
 }
